// src/compra/dto/create-compra.dto.ts
export class CreateCompraDto {
  idTrabajador: string;
  idArticulo: string;
  idMetodopago: string;
  idOCasionCompra: string;
  idMetodoEntrega: string;
  unidades: string;
  comision: string;
  fecha: string;
  idCliente: string;
  entregaCalleNum: string;
  entregaCol: string;
  entregaMun: string;
  entregaEdo: string;
  entregaCp: string;
}